split OAuth token shape into dedicated core microcrate
