+++ Cargo.toml
@@ -20,6 +20,7 @@ members = [
   "crates/uselesskey-core-sink",
   "crates/uselesskey-core-token",
   "crates/uselesskey-core-token-shape",
+  "crates/uselesskey-core-oauth-shape",
   "crates/uselesskey-core-jwk-builder",
   "crates/uselesskey-core-jwks-order",
   "crates/uselesskey-core-jwk",
@@ -75,6 +76,7 @@ uselesskey-core-factory = { path = "crates/uselesskey-core-factory", version = "
 uselesskey-core-jwks-order = { path = "crates/uselesskey-core-jwks-order", version = "0.1.0" }
 uselesskey-core-hash = { path = "crates/uselesskey-core-hash", version = "0.1.0", default-features = false }
 uselesskey-core-token-shape = { path = "crates/uselesskey-core-token-shape", version = "0.1.0" }
+uselesskey-core-oauth-shape = { path = "crates/uselesskey-core-oauth-shape", version = "0.1.0" }
 uselesskey-core-id = { path = "crates/uselesskey-core-id", version = "0.1.0", default-features = false }
 uselesskey-core-seed = { path = "crates/uselesskey-core-seed", version = "0.1.0", default-features = false }
 uselesskey-core-kid = { path = "crates/uselesskey-core-kid", version = "0.1.0" }
