+++ crates/uselesskey-core-token-shape/README.md
@@ -8,5 +8,6 @@ Low-level token shape primitives for `uselesskey`.
 - Generate opaque bearer token shapes.
 - Generate OAuth-like JWT-access-token shapes without signing.
 
-This crate intentionally contains only token-shape construction and is used by
-`uselesskey-core-token` and higher-level token fixture crates.
+This crate intentionally contains token-shape composition logic for API keys
+and bearer tokens, and re-exports OAuth JWT-shape generation from
+`uselesskey-core-oauth-shape`.
