+++ crates/uselesskey-core-token-shape/src/lib.rs
@@ -31,9 +31,10 @@
 use base64::Engine as _;
 use base64::engine::general_purpose::URL_SAFE_NO_PAD;
 use rand_core::RngCore;
-
-use serde_json::json;
 pub use uselesskey_core_base62::random_base62;
+pub use uselesskey_core_oauth_shape::{
+    OAUTH_JTI_BYTES, OAUTH_SIGNATURE_BYTES, generate_oauth_access_token,
+};
 
 /// Prefix used for API-key token fixtures.
 pub const API_KEY_PREFIX: &str = "uk_test_";
@@ -43,12 +44,6 @@ pub const API_KEY_RANDOM_LEN: usize = 32;
 
 /// Number of raw random bytes in opaque bearer tokens.
 pub const BEARER_RANDOM_BYTES: usize = 32;
-
-/// Number of random bytes used for OAuth `jti`.
-pub const OAUTH_JTI_BYTES: usize = 16;
-
-/// Number of random bytes used for OAuth signature-like segment.
-pub const OAUTH_SIGNATURE_BYTES: usize = 32;
 
 /// Token shape kind.
 pub use uselesskey_token_spec::TokenSpec as TokenKind;
@@ -79,31 +74,6 @@ pub fn generate_bearer_token(rng: &mut impl RngCore) -> String {
     let mut bytes = [0u8; BEARER_RANDOM_BYTES];
     rng.fill_bytes(&mut bytes);
     URL_SAFE_NO_PAD.encode(bytes)
-}
-
-/// Generate an OAuth access token fixture in JWT shape (`header.payload.signature`).
-pub fn generate_oauth_access_token(label: &str, rng: &mut impl RngCore) -> String {
-    let header = URL_SAFE_NO_PAD.encode(r#"{"alg":"RS256","typ":"JWT"}"#);
-
-    let mut jti = [0u8; OAUTH_JTI_BYTES];
-    rng.fill_bytes(&mut jti);
-
-    let payload = json!({
-        "iss": "uselesskey",
-        "sub": label,
-        "aud": "tests",
-        "scope": "fixture.read",
-        "jti": URL_SAFE_NO_PAD.encode(jti),
-        "exp": 2_000_000_000u64,
-    });
-    let payload_json = serde_json::to_vec(&payload).expect("payload JSON");
-    let payload_segment = URL_SAFE_NO_PAD.encode(payload_json);
-
-    let mut signature = [0u8; OAUTH_SIGNATURE_BYTES];
-    rng.fill_bytes(&mut signature);
-    let signature_segment = URL_SAFE_NO_PAD.encode(signature);
-
-    format!("{header}.{payload_segment}.{signature_segment}")
 }
 
 #[cfg(test)]
